requirements.txt
confluent-kafka
fmi-weather-client
google-cloud-bigquery
pyyaml
